Split entitlement types
