+++ packages/types/src/index.ts
@@ -3,9 +3,10 @@ export * from './discord/applications.js'
 export * from './discord/auditLogs.js'
 export * from './discord/autoModeration.js'
 export * from './discord/channels.js'
+export * from './discord/emojis.js'
+export * from './discord/entitlements.js'
 export * from './discord/gateway.js'
 export * from './discord/interactions.js'
-export * from './discord/emojis.js'
 export * from './discord/reference.js'
 export * from './discord/webhookEvents.js'
 export * from './discordeno.js'
